// src/app/viaturas/models/Viaturas.model.ts
export interface Viaturas {
  id: number;
  placa: string;
  renavan: number;
  chassi: string;
  modelo: string;
  combustivel: string;
  capacidadePassageiros: number;
  quantidadeDePortas: number;
  capacidadeDoTanque: number;
  autonomia: number;
  utilizacao: string;
  statusViatura: string;
  nee: string;
  eb: string;
}